apps/python-renderer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
supabase==2.0.0
redis==5.0.1
websockets==12.0

